packages/argus_cli/pyproject.toml
[build-system]
requires = ["hatchling>=1.20"]
build-backend = "hatchling.build"

[project]
name = "argus-cli"
version = "0.1.0"
description = "Interactive CLI for Argus MCP — dual-mode REPL + one-shot commands"
readme = "README.md"
license = "GPL-3.0-only"
requires-python = ">=3.10"
authors = [{ name = "Daniel Diaz Santiago" }]
keywords = ["argus", "mcp", "cli", "repl"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Typing :: Typed",
]
dependencies = [
    "argus-mcp",
    "typer>=0.15.0",
    "rich>=14.0.0",
    "prompt-toolkit>=3.0.50",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.1.0",
    "httpx-sse>=0.4.0",
    "shellingham>=1.5.4",
    "tabulate>=0.9",
    "textual>=1.0.0",
    "textual-plotext>=1.0.0",
    "typing-extensions>=4.12.0; python_version < '3.11'",
]

[project.optional-dependencies]
# tui deps (textual, textual-plotext) moved to core — argus-tui entry point requires them
tui = []
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0",
    "ruff>=0.11.0",
    "mypy>=1.15",
]

## Entry points
[project.scripts]
argus = "argus_cli.main:app"
argus-cli = "argus_cli.main:app"
argus-tui = "argus_cli.tui:launch"

[project.urls]
Repository = "https://github.com/diaz3618/argus-mcp"

# ── Ruff ────────────────────────────────────────────────────────────────

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "N",
    "UP",
    "B",
    "S",
    "T20",
    "SIM",
    "TC",
    "RUF",
]
ignore = [
    "T201",  # Allow print() — CLI tool
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "TC"]
"argus_cli/commands/**" = ["TC"]
"argus_cli/client.py" = ["TC"]
"argus_cli/output.py" = ["TC"]
"argus_cli/repl/**" = ["TC"]
"argus_cli/widgets/**" = ["TC"]
# TUI files copied from server package — will be refactored in Phase 2+
"argus_cli/tui/**" = ["E501", "RUF003", "RUF012", "S110", "S607", "SIM", "TC", "UP045"]

[tool.ruff.lint.isort]
known-first-party = ["argus_cli"]

# ── Mypy ────────────────────────────────────────────────────────────────

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_untyped_calls = false
disallow_incomplete_defs = false

# ── Pytest ──────────────────────────────────────────────────────────────

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["argus_cli"]
branch = true

[tool.coverage.report]
fail_under = 80
show_missing = true
skip_covered = true
skip_empty = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__",
    "@overload",
    "raise NotImplementedError",
]
